[locale] Add Kurmanji translations and move Central Kurdish to ckb.js
- About the ckb rename: The reason for the rename is that the wrong code was picked for Central Kurdish when it was added first.

Related discussions: https://github.com/moment/moment/pull/5895, https://github.com/moment/moment/pull/5968
